services/ai/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
redis==5.2.1
scikit-learn==1.6.1
xgboost==2.1.3
pandas==2.2.3
numpy==2.2.1
langchain==0.3.14
langchain-openai==0.3.0
langchain-anthropic==0.3.4
python-dotenv==1.0.1
httpx==0.28.1
prometheus-client==0.22.0
structlog==25.1.0
